# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Track Leader API",
    "description": "API for Track Leader - Activity tracking and segment leaderboards",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/me/achievements": {
      "get": {
        "tags": [
          "achievements"
        ],
        "summary": "Get the authenticated user's achievements.",
        "operationId": "get_my_achievements",
        "parameters": [
          {
            "name": "include_lost",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User achievements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementWithSegment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/segments/{id}": {
      "get": {
        "tags": [
          "segments"
        ],
        "operationId": "get_segment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Segment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Segment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Segment"
                }
              }
            }
          },
          "404": {
            "description": "Segment not found"
          }
        }
      }
    },
    "/segments/{id}/achievements": {
      "get": {
        "tags": [
          "achievements"
        ],
        "summary": "Get current achievement holders for a segment.",
        "operationId": "get_segment_achievements",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Segment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Segment achievement holders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentAchievements"
                }
              }
            }
          },
          "404": {
            "description": "Segment not found"
          }
        }
      }
    },
...
    "/users/{id}/achievements": {
      "get": {
        "tags": [
          "achievements"
        ],
        "operationId": "get_user_achievements",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "include_lost",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User achievements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementWithSegment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AchievementHolder": {
        "type": "object",
        "description": "Holder of an achievement with their details",
        "required": [
          "user_id",
          "user_name",
          "achievement_type",
          "earned_at"
        ],
        "properties": {
          "achievement_type": {
            "$ref": "#/components/schemas/AchievementType"
          },
          "earned_at": {
            "type": "string",
            "format": "date-time"
          },
          "effort_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "elapsed_time_seconds": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "AchievementType": {
        "type": "string",
        "description": "Type of achievement/crown",
        "enum": [
          "kom",
          "qom",
          "course_record"
        ]
      },
      "AchievementWithSegment": {
        "type": "object",
        "description": "Achievement with segment details for display",
        "required": [
          "id",
          "user_id",
          "segment_id",
          "achievement_type",
          "earned_at",
          "segment_name",
          "segment_distance_meters",
          "segment_activity_type_id"
        ],
        "properties": {
          "achievement_type": {
            "$ref": "#/components/schemas/AchievementType"
          },
          "earned_at": {
            "type": "string",
            "format": "date-time"
          },
          "effort_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "effort_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lost_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "segment_activity_type_id": {
            "type": "string",
            "format": "uuid"
          },
          "segment_distance_meters": {
            "type": "number",
            "format": "double"
          },
          "segment_id": {
            "type": "string",
            "format": "uuid"
          },
          "segment_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Segment": {
        "type": "object",
        "required": [
          "id",
          "creator_id",
          "name",
          "activity_type_id",
          "distance_meters",
          "visibility",
          "created_at"
        ],
        "properties": {
          "activity_type_id": {
            "type": "string",
            "format": "uuid"
          },
          "average_grade": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "climb_category": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "creator_id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "distance_meters": {
            "type": "number",
            "format": "double"
          },
          "elevation_gain_meters": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "elevation_loss_meters": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "max_grade": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          }
        }
      },
      "SegmentAchievements": {
        "type": "object",
        "description": "Current achievement holders for a segment",
        "required": [
          "segment_id"
        ],
        "properties": {
          "kom": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AchievementHolder"
              }
            ]
          },
          "qom": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AchievementHolder"
              }
            ]
          },
          "segment_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}